requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
eventlet==0.33.0
python-dotenv==0.19.0
requests==2.26.0
websocket-client==1.2.1
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
isort==5.9.3
flask-cors==3.0.10
cryptography==41.0.1
fyers-apiv3==0.0.1
gunicorn==20.1.0 